Done with step 1, 2, 3, 4. Summary page and form handling with save d…

+++ src/main/java/pl/coderslab/charity/donation/DonationService.java
@@ -18,5 +18,5 @@ public interface DonationService {
 
     Integer bagsQuantity();
 
-    Integer donationQuantity();
+    Long donationQuantity();
 }
